Reorganise contract decoders (#1138)
This reorganises contract decoders to their relevant domains, with subfolders for their tests so that they are not included in builds.

---

* Reorganise contract decoders

* Move tests and builders to relevant folders

* Remove duplicate file

* Remove proxy factory tests
