+++ src/domain/relay/contracts/proxy-factory-decoder.helper.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { parseAbi } from 'viem';
-import { AbiDecoder } from '@/domain/alerts/contracts/abi-decoder.helper';
+import { AbiDecoder } from '@/domain/contracts/contracts/abi-decoder.helper';
 
 const PROXY_FACTORY_ABI = parseAbi([
   'function createProxyWithNonce(address _singleton, bytes memory initializer, uint256 saltNonce)',
