+++ src/domain/relay/relay.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { LimitAddressesMapper } from '@/domain/relay/limit-addresses.mapper';
 import { Erc20ContractHelper } from '@/domain/relay/contracts/erc20-contract.helper';
 import { SafeContractHelper } from '@/domain/relay/contracts/safe-contract.helper';
-import { MultiSendDecoder } from '@/domain/alerts/contracts/multi-send-decoder.helper';
+import { MultiSendDecoder } from '@/domain/contracts/contracts/multi-send-decoder.helper';
 import { ProxyFactoryDecoder } from '@/domain/relay/contracts/proxy-factory-decoder.helper';
 
 @Module({
